fix(dishwasher): make singleton work
Using a static variable to store the singleton instance
does not work as we need to access it using `BalayDishwasher.instance`.
This lead esbuild to transpile the code to
`var BalayDishwasher = class _BalayDishwasher ...` which is then reflect
in the store key, and is not the same key used in the instance and the
Constructor.